python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Jonathon Broughton <jonathon@stardotbmp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
specklepy = "^2.12.0"
Pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
